Addressed review comments

+++ src/tempShipment/controller.js
@@ -37,12 +37,17 @@ async function createOne(req, res) {
 }
 
 /**
- * Create a new temp shipment given a manifest
+ * Delete a new temp shipment given a manifest
  */
 async function deleteOne(req, res) {
   ExpressHandler(
     async () => {
       const { tsid } = req.params;
+
+      const tmpShipment = await TempShipment.findById(tsid);
+
+      if (!tmpShipment)
+        return HTTPError("Temporary Shipment does not exist", 400);
 
       // delete temp shipment
       await TempShipment.deleteOne({ _id: tsid });
